refactor: extract Rive init/crash-guard into RiveInitializer.kt
Move the Rive crash-loop guard and SDK initialization (isRiveInitialized,
resetRiveInitialized, riveGuardPrefs, initializeRive and their constants)
out of VoltixApplication.kt into a dedicated RiveInitializer.kt in the same
package. Pure code move — no behavior change; the Application class keeps
calling initializeRive(this). Update the test's mockkStatic facade target
from VoltixApplicationKt to RiveInitializerKt.
